Support handling cases where multiple integration records found for the same file in source depot
## Issue:

In case when source depot contains multiple integration records of the same depotfile, the original script only takes the first record.

Such cases can be easily observed on dirty moves, that in **one CL**, file A is branched into file B and file A is deleted. Running `p4 filelog` with file B's path will add into the result the older integration record of A.

## Solution:

Preprocess the filelog result so that only the latest record for each file as returned from `p4 filelog` is being processed.